Add legacy adapter
