+++ src/components/Settings/EventLog/EventLog.tsx
@@ -1,63 +1,43 @@
 import styled from "@emotion/styled";
-import Badge, { Variant } from "@leafygreen-ui/badge";
 import Button from "@leafygreen-ui/button";
 import Card from "@leafygreen-ui/card";
-import { Table, TableHeader, Row, Cell } from "@leafygreen-ui/table";
-import { fontFamilies } from "@leafygreen-ui/tokens";
 import { Subtitle } from "@leafygreen-ui/typography";
 import { size } from "constants/tokens";
-import { EventDiffLine, EventValue, getEventDiffLines } from "./EventLogDiffs";
+import { EventDiffTable } from "./EventDiffTable";
 import { Header } from "./Header";
+import { Event } from "./types";
 
-export const EventLog = ({ allEventsFetched, events, handleFetchMore }) => {
+type EventLogProps = {
+  allEventsFetched: boolean;
+  eventRenderer?: (event: Event) => React.ReactNode;
+  events: Event[];
+  handleFetchMore: () => void;
+};
+
+export const EventLog: React.FC<EventLogProps> = ({
+  allEventsFetched,
+  eventRenderer,
+  events,
+  handleFetchMore,
+}) => {
   const allEventsFetchedCopy =
     events.length > 0 ? "No more events to show." : "No events to show.";
 
   return (
     <Container data-cy="event-log">
-      {events.map(({ after, before, timestamp, user }) => (
-        <EventLogCard key={`event_log_${timestamp}`} data-cy="event-log-card">
-          <Header user={user} timestamp={timestamp} />
-          <Table
-            data={getEventDiffLines(before, after)}
-            columns={[
-              <TableHeader
-                key="key"
-                label="Property"
-                sortBy={(datum: EventDiffLine) => datum.key}
-              />,
-              <TableHeader
-                key="before"
-                label="Before"
-                sortBy={(datum: EventDiffLine) => JSON.stringify(datum.before)}
-              />,
-              <TableHeader
-                key="after"
-                label="After"
-                sortBy={(datum: EventDiffLine) => JSON.stringify(datum.after)}
-              />,
-            ]}
-          >
-            {({ datum }) => (
-              <Row key={datum.key} data-cy="event-log-table-row">
-                <Cell>
-                  <CellText>{datum.key}</CellText>
-                </Cell>
-                <Cell>
-                  <CellText>{renderEventValue(datum.before)}</CellText>
-                </Cell>
-                <Cell>
-                  {renderEventValue(datum.after) === null ? (
-                    <Badge variant={Variant.Red}>Deleted</Badge>
-                  ) : (
-                    <CellText>{renderEventValue(datum.after)}</CellText>
-                  )}
-                </Cell>
-              </Row>
+      {events.map((event) => {
+        const { after, before, timestamp, user } = event;
+        return (
+          <EventLogCard key={`event_log_${timestamp}`} data-cy="event-log-card">
+            <Header user={user} timestamp={timestamp} />
+            {eventRenderer ? (
+              eventRenderer(event)
+            ) : (
+              <EventDiffTable after={after} before={before} />
             )}
-          </Table>
-        </EventLogCard>
-      ))}
+          </EventLogCard>
+        );
+      })}
       {!allEventsFetched && !!events.length && (
         <Button
           data-cy="load-more-button"
@@ -84,33 +64,3 @@ const EventLogCard = styled(Card)`
   margin-bottom: ${size.l};
   padding: ${size.m};
 `;
-
-const CellText = styled.span`
-  font-family: ${fontFamilies.code};
-  font-size: 12px;
-  line-height: 16px;
-  word-break: break-all;
-`;
-
-const renderEventValue = (value: EventValue): string => {
-  if (value === null || value === undefined) {
-    return null;
-  }
-  if (typeof value === "boolean") {
-    return String(value);
-  }
-
-  if (typeof value === "string") {
-    return `"${value}"`;
-  }
-
-  if (typeof value === "number") {
-    return value;
-  }
-
-  if (Array.isArray(value)) {
-    return JSON.stringify(value).replaceAll(",", ",\n");
-  }
-
-  return JSON.stringify(value);
-};
