+++ src/components/Settings/EventLog/index.ts
@@ -1,2 +1,3 @@
+export { EventDiffTable } from "./EventDiffTable";
 export { EventLog } from "./EventLog";
 export { EVENT_LIMIT, useEvents } from "./useEvents";
