+++ src/pages/distroSettings/tabs/EventLogTab/EventLogTab.tsx
@@ -1,5 +1,6 @@
 import { useParams } from "react-router-dom";
-import { EventLog } from "components/Settings/EventLog";
+import { EventDiffTable, EventLog } from "components/Settings/EventLog";
+import { LegacyEventEntry } from "./LegacyEventEntry";
 import { useDistroEvents } from "./useDistroEvents";
 
 type TabProps = {
@@ -22,6 +23,13 @@ export const EventLogTab: React.FC<TabProps> = ({ limit }) => {
     <EventLog
       allEventsFetched={allEventsFetched}
       events={events}
+      eventRenderer={({ after, before, data }) =>
+        after && before ? (
+          <EventDiffTable after={after} before={before} />
+        ) : (
+          <LegacyEventEntry data={data} />
+        )
+      }
       handleFetchMore={() => {
         fetchMore({
           variables: {
